+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * @iov/bcp: Add `AtomicSwapMerger`
 * @iov/bcp: Add `signedBy` field to `BcpTxQuery`.
+* @iov/keycontrol: Add `Keyring.getAllIdentities`
 
 Breaking changes